Moved graph classes into their own namespace, many small cleanups along the way.

+++ src/NGenerics.Tests/DataStructures/Trees/BinarySearchTreeTests/Remove.cs
@@ -48,7 +48,7 @@ namespace NGenerics.Tests.DataStructures.Trees.BinarySearchTreeTests
             {
                 while (enumerator.MoveNext())
                 {
-                    Assert.IsTrue(tree.Remove(enumerator.Current));
+                    Assert.IsTrue(tree.Remove((int) enumerator.Current));
                 }
             }
         }
